+++ python/python/glide/async_commands/transaction.py
@@ -2534,6 +2534,45 @@ class BaseTransaction:
 
         return self.append_command(RequestType.XAutoClaim, args)
 
+    def xinfo_groups(
+        self: TTransaction,
+        key: TEncodable,
+    ) -> TTransaction:
+        """
+        Returns the list of all consumer groups and their attributes for the stream stored at `key`.
+
+        See https://valkey.io/commands/xinfo-groups for more details.
+
+        Args:
+            key (TEncodable): The key of the stream.
+
+        Command response:
+            List[Mapping[bytes, Union[bytes, int, None]]]: A list of mappings, where each mapping represents the
+                attributes of a consumer group for the stream at `key`.
+        """
+        return self.append_command(RequestType.XInfoGroups, [key])
+
+    def xinfo_consumers(
+        self: TTransaction,
+        key: TEncodable,
+        group_name: TEncodable,
+    ) -> TTransaction:
+        """
+        Returns the list of all consumers and their attributes for the given consumer group of the stream stored at
+        `key`.
+
+        See https://valkey.io/commands/xinfo-consumers for more details.
+
+        Args:
+            key (TEncodable): The key of the stream.
+            group_name (TEncodable): The consumer group name.
+
+        Command response:
+            List[Mapping[bytes, Union[bytes, int]]]: A list of mappings, where each mapping contains the attributes of a
+                consumer for the given consumer group of the stream at `key`.
+        """
+        return self.append_command(RequestType.XInfoConsumers, [key, group_name])
+
     def geoadd(
         self: TTransaction,
         key: TEncodable,
@@ -4492,7 +4531,7 @@ class BaseTransaction:
                                      in the list. A `max_len` of 0 means unlimited amount of comparisons.
 
         Command Response:
-            Union[int, list[int], None]: The index of the first occurrence of `element`,
+            Union[int, List[int], None]: The index of the first occurrence of `element`,
             or None if `element` is not in the list.
             With the `count` option, a list of indices of matching elements will be returned.