+++ python/python/tests/test_async_client.py
@@ -6,9 +6,8 @@ import asyncio
 import copy
 import math
 import time
-from collections.abc import Mapping
 from datetime import date, datetime, timedelta, timezone
-from typing import Any, Dict, List, Tuple, Union, cast
+from typing import Any, Dict, List, Mapping, Tuple, Union, cast
 
 import pytest
 from glide import ClosingError, RequestError, Script
@@ -6355,6 +6354,189 @@ class TestCommands:
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_xinfo_groups_xinfo_consumers(
+        self, redis_client: TGlideClient, protocol
+    ):
+        key = get_random_string(10)
+        group_name1 = get_random_string(10)
+        group_name2 = get_random_string(10)
+        consumer1 = get_random_string(10)
+        consumer2 = get_random_string(10)
+        stream_id0_0 = "0-0"
+        stream_id1_0 = "1-0"
+        stream_id1_1 = "1-1"
+        stream_id1_2 = "1-2"
+        stream_id1_3 = "1-3"
+
+        # setup: add 3 entries to stream, create consumer group and consumer1, read 1 entry from stream with consumer1
+        assert (
+            await redis_client.xadd(
+                key, [("f1", "v1"), ("f2", "v2")], StreamAddOptions(stream_id1_0)
+            )
+            == stream_id1_0.encode()
+        )
+        assert (
+            await redis_client.xadd(key, [("f3", "v3")], StreamAddOptions(stream_id1_1))
+            == stream_id1_1.encode()
+        )
+        assert (
+            await redis_client.xadd(key, [("f4", "v4")], StreamAddOptions(stream_id1_2))
+            == stream_id1_2.encode()
+        )
+        assert await redis_client.xgroup_create(key, group_name1, stream_id0_0) == OK
+        assert await redis_client.xreadgroup(
+            {key: ">"}, group_name1, consumer1, StreamReadGroupOptions(count=1)
+        ) == {key.encode(): {stream_id1_0.encode(): [[b"f1", b"v1"], [b"f2", b"v2"]]}}
+
+        # sleep to ensure the idle time value and inactive time value returned by xinfo_consumers is > 0
+        time.sleep(2)
+        consumers_result = await redis_client.xinfo_consumers(key, group_name1)
+        assert len(consumers_result) == 1
+        consumer1_info = consumers_result[0]
+        assert consumer1_info.get(b"name") == consumer1.encode()
+        assert consumer1_info.get(b"pending") == 1
+        assert cast(int, consumer1_info.get(b"idle")) > 0
+        if not await check_if_server_version_lt(redis_client, "7.2.0"):
+            assert (
+                cast(int, consumer1_info.get(b"inactive"))
+                > 0  # "inactive" was added in Redis 7.2.0
+            )
+
+        # create consumer2 and read the rest of the entries with it
+        assert (
+            await redis_client.xgroup_create_consumer(key, group_name1, consumer2)
+            is True
+        )
+        assert await redis_client.xreadgroup({key: ">"}, group_name1, consumer2) == {
+            key.encode(): {
+                stream_id1_1.encode(): [[b"f3", b"v3"]],
+                stream_id1_2.encode(): [[b"f4", b"v4"]],
+            }
+        }
+
+        # verify that xinfo_consumers contains info for 2 consumers now
+        # test with byte string args
+        consumers_result = await redis_client.xinfo_consumers(
+            key.encode(), group_name1.encode()
+        )
+        assert len(consumers_result) == 2
+
+        # add one more entry
+        assert (
+            await redis_client.xadd(key, [("f5", "v5")], StreamAddOptions(stream_id1_3))
+            == stream_id1_3.encode()
+        )
+
+        groups = await redis_client.xinfo_groups(key)
+        assert len(groups) == 1
+        group1_info = groups[0]
+        assert group1_info.get(b"name") == group_name1.encode()
+        assert group1_info.get(b"consumers") == 2
+        assert group1_info.get(b"pending") == 3
+        assert group1_info.get(b"last-delivered-id") == stream_id1_2.encode()
+        if not await check_if_server_version_lt(redis_client, "7.0.0"):
+            assert (
+                group1_info.get(b"entries-read")
+                == 3  # we have read stream entries 1-0, 1-1, and 1-2
+            )
+            assert (
+                group1_info.get(b"lag")
+                == 1  # we still have not read one entry in the stream, entry 1-3
+            )
+
+        # verify xgroup_set_id effects the returned value from xinfo_groups
+        assert await redis_client.xgroup_set_id(key, group_name1, stream_id1_1) == OK
+        # test with byte string arg
+        groups = await redis_client.xinfo_groups(key.encode())
+        assert len(groups) == 1
+        group1_info = groups[0]
+        assert group1_info.get(b"name") == group_name1.encode()
+        assert group1_info.get(b"consumers") == 2
+        assert group1_info.get(b"pending") == 3
+        assert group1_info.get(b"last-delivered-id") == stream_id1_1.encode()
+        # entries-read and lag were added to the result in 7.0.0
+        if not await check_if_server_version_lt(redis_client, "7.0.0"):
+            assert (
+                group1_info.get(b"entries-read")
+                is None  # gets set to None when we change the last delivered ID
+            )
+            assert (
+                group1_info.get(b"lag")
+                is None  # gets set to None when we change the last delivered ID
+            )
+
+        if not await check_if_server_version_lt(redis_client, "7.0.0"):
+            # verify xgroup_set_id with entries_read_id effects the returned value from xinfo_groups
+            assert (
+                await redis_client.xgroup_set_id(
+                    key, group_name1, stream_id1_1, entries_read_id="1"
+                )
+                == OK
+            )
+            groups = await redis_client.xinfo_groups(key)
+            assert len(groups) == 1
+            group1_info = groups[0]
+            assert group1_info.get(b"name") == group_name1.encode()
+            assert group1_info.get(b"consumers") == 2
+            assert group1_info.get(b"pending") == 3
+            assert group1_info.get(b"last-delivered-id") == stream_id1_1.encode()
+            assert group1_info.get(b"entries-read") == 1
+            assert (
+                group1_info.get(b"lag")
+                == 3  # lag is calculated as number of stream entries minus entries-read
+            )
+
+        # add one more consumer group
+        assert await redis_client.xgroup_create(key, group_name2, stream_id0_0) == OK
+
+        # verify that xinfo_groups contains info for 2 consumer groups now
+        groups = await redis_client.xinfo_groups(key)
+        assert len(groups) == 2
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_xinfo_groups_xinfo_consumers_edge_cases_and_failures(
+        self, redis_client: TGlideClient, protocol
+    ):
+        key = get_random_string(10)
+        string_key = get_random_string(10)
+        non_existing_key = get_random_string(10)
+        group_name = get_random_string(10)
+        stream_id1_0 = "1-0"
+
+        # passing a non-existing key raises an error
+        with pytest.raises(RequestError):
+            await redis_client.xinfo_groups(non_existing_key)
+        with pytest.raises(RequestError):
+            await redis_client.xinfo_consumers(non_existing_key, group_name)
+
+        assert (
+            await redis_client.xadd(
+                key, [("f1", "v1"), ("f2", "v2")], StreamAddOptions(stream_id1_0)
+            )
+            == stream_id1_0.encode()
+        )
+
+        # passing a non-existing group raises an error
+        with pytest.raises(RequestError):
+            await redis_client.xinfo_consumers(key, "non_existing_group")
+
+        # no groups exist yet
+        assert await redis_client.xinfo_groups(key) == []
+
+        assert await redis_client.xgroup_create(key, group_name, stream_id1_0) == OK
+        # no consumers exist yet
+        assert await redis_client.xinfo_consumers(key, group_name) == []
+
+        # key exists, but it is not a stream
+        assert await redis_client.set(string_key, "foo") == OK
+        with pytest.raises(RequestError):
+            await redis_client.xinfo_groups(string_key)
+        with pytest.raises(RequestError):
+            await redis_client.xinfo_consumers(string_key, group_name)
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_xgroup_set_id(
         self, redis_client: TGlideClient, cluster_mode, protocol, request
     ):
@@ -7297,7 +7479,7 @@ class TestCommands:
         await redis_client.function_load(code)
 
         check_function_list_response(
-            await redis_client.function_list(lib_name),
+            await redis_client.function_list(lib_name.encode()),
             lib_name,
             {func_name: None},
             {func_name: {b"no-writes"}},
@@ -7319,7 +7501,9 @@ class TestCommands:
         )
 
         no_args_response = await redis_client.function_list()
-        wildcard_pattern_response = await redis_client.function_list("*", False)
+        wildcard_pattern_response = await redis_client.function_list(
+            "*".encode(), False
+        )
         assert len(no_args_response) == original_functions_count + 1
         assert len(wildcard_pattern_response) == original_functions_count + 1
         check_function_list_response(
@@ -8489,7 +8673,7 @@ class TestMultiKeyCommandCrossSlot:
     async def test_multi_key_command_returns_cross_slot_error(
         self, redis_client: GlideClusterClient
     ):
-        promises: list[Any] = [
+        promises: List[Any] = [
             redis_client.blpop(["abc", "zxy", "lkn"], 0.1),
             redis_client.brpop(["abc", "zxy", "lkn"], 0.1),
             redis_client.rename("abc", "zxy"),
@@ -9028,7 +9212,7 @@ class TestClusterRoutes:
             ),
         )
         result_cursor = str(result[result_cursor_index])
-        result_iteration_collection: dict[str, str] = convert_list_to_dict(
+        result_iteration_collection: Dict[str, str] = convert_list_to_dict(
             result[result_collection_index]
         )
         full_result_map.update(result_iteration_collection)
@@ -9052,7 +9236,8 @@ class TestClusterRoutes:
             full_result_map.update(next_result_collection)
             result_iteration_collection = next_result_collection
             result_cursor = next_result_cursor
-        assert (num_map_with_str_scores | char_map_with_str_scores) == full_result_map
+        num_map_with_str_scores.update(char_map_with_str_scores)
+        assert num_map_with_str_scores == full_result_map
 
         # Test match pattern
         result = await redis_client.zscan(key1, initial_cursor, match="*")
@@ -9140,7 +9325,7 @@ class TestClusterRoutes:
             ),
         )
         result_cursor = str(result[result_cursor_index])
-        result_iteration_collection: dict[str, str] = convert_list_to_dict(
+        result_iteration_collection: Dict[str, str] = convert_list_to_dict(
             result[result_collection_index]
         )
         full_result_map.update(result_iteration_collection)
@@ -9164,7 +9349,8 @@ class TestClusterRoutes:
             full_result_map.update(next_result_collection)
             result_iteration_collection = next_result_collection
             result_cursor = next_result_cursor
-        assert (num_map | char_map) == full_result_map
+        num_map.update(char_map)
+        assert num_map == full_result_map
 
         # Test match pattern
         result = await redis_client.hscan(key1, initial_cursor, match="*")